Moved cloudcmd into an external function
